Refactored and moved actual formula generation to its own class
