+++ CompatableExcelCleaner/CompatableExcelCleaner.csproj
@@ -63,11 +63,13 @@
     <Compile Include="AbstractMergeCleaner.cs" />
     <Compile Include="BackupMergeCleaner.cs" />
     <Compile Include="FileCleaner.cs" />
-    <Compile Include="FormulaMaker.cs" />
+    <Compile Include="FormulaManager.cs" />
+    <Compile Include="IFormulaGenerator.cs" />
     <Compile Include="IMergeCleaner.cs" />
     <Compile Include="MergeType.cs" />
     <Compile Include="PrimaryMergeCleaner.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RowSegmentFormulaGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
