+++ CompatableExcelCleaner/FormulaManager.cs
@@ -9,7 +9,7 @@ namespace CompatableExcelCleaner
     /// <summary>
     /// Replaces static values in excel files with formulas that will change when the data is updated
     /// </summary>
-    public class FormulaMaker
+    public class FormulaManager
     {
 
 
@@ -19,7 +19,7 @@ namespace CompatableExcelCleaner
 
 
 
-        static FormulaMaker() 
+        static FormulaManager() 
         {
 
             //Fill our dictionary with all the reports and all the data we need to give them formulas
@@ -104,17 +104,10 @@ namespace CompatableExcelCleaner
                 for (int i = 0; i < package.Workbook.Worksheets.Count; i++)
                 {
                     worksheet = package.Workbook.Worksheets[i];
-                    
-                    foreach (string header in headers)              //for each header in the report that needs a formula 
-                    {
-                        var ranges = GetRowRangeForFormula(worksheet, header);
 
-                        foreach (var item in ranges)                // for each instance of that header
-                        {
-                            FillInFormulas(worksheet, item.Item1, item.Item2, item.Item3);
-                        }
-                    }
-
+                    //call formula generator
+                    IFormulaGenerator formulaGenerator = new RowSegmentFormulaGenerator();
+                    formulaGenerator.InsertFormulas(worksheet, headers);
                 }
 
 
@@ -125,108 +118,9 @@ namespace CompatableExcelCleaner
 
 
 
-        /// <summary>
-        /// Gets the row numbers of the first and last rows that should be included in the formula
-        /// </summary>
-        /// <param name="worksheet">the worksheet currently being given formulas</param>
-        /// <param name="targetText">the text to look for to signal the start and end row</param>
-        /// <returns>a tuple containing the start-row, end-row, and column of the formula range</returns>
-        private static IEnumerable<Tuple<int, int, int>> GetRowRangeForFormula(ExcelWorksheet worksheet, string targetText)
-        {
-            ExcelRange cell;
 
 
-            for(int row = 1; row < worksheet.Dimension.End.Row; row++)
-            {
-                for(int col = 1; col < worksheet.Dimension.End.Column; col++)
-                {
-                    cell = worksheet.Cells[row, col];
-
-                    if (cell.Text == targetText)
-                    {
-                        //search for end of sequence
-                        int end = FindEndOfFormulaRange(worksheet, row, col, "Total " + targetText);
-
-                        if (end > 0)
-                        {
-                            yield return new Tuple<int, int, int>(row, end, col);
-
-                            //for the next iteration, jump to after the formula range we just returned
-                            row = end + 1;
-                            col = 1;
-                        }
-                    }
-                }
-            }
-        }
-
-
-
-
-        /// <summary>
-        /// Given the Cell coordinates of the starting cell in a formula range, finds the ending cell for 
-        /// that range.
-        /// </summary>
-        /// <param name="worksheet">the worksheet currently being given formulas</param>
-        /// <param name="row">the row number of the starting cell in the formula range</param>
-        /// <param name="col">the column number of the starting cell in the formula range</param>
-        /// <param name="targetText">the text to look for that signals the end cell of the formula range</param>
-        /// <returns>the row number of the last cell in the formula range, or -1 if no appropriate last cell is found</returns>
-        private static int FindEndOfFormulaRange(ExcelWorksheet worksheet, int row, int col, string targetText)
-        {
-            ExcelRange cell;
-
-            for(int i = row + 1; i < worksheet.Dimension.End.Row; i++)
-            {
-                cell = worksheet.Cells[i, col];
-                if (cell.Text == targetText)
-                {
-                    return i;
-                }
-            }
-
-
-            return -1;
-        }
-
-
-
-
-        /// <summary>
-        /// Inserts the formulas in each cell in the formula range that requires it.
-        /// </summary>
-        /// <param name="worksheet">the worksheet currently being given formulas</param>
-        /// <param name="startRow">the first row of the formula range (containing the header)</param>
-        /// <param name="endRow">the last row of the formula range (containing the total)</param>
-        /// <param name="col">the column of the header and total for the formula range</param>
-        private static void FillInFormulas(ExcelWorksheet worksheet, int startRow, int endRow, int col)
-        {
-
-            ExcelRange cell;
-            
-
-
-            //Often there are multiple columns that require a formula, so we need to iterate
-            //and apply the formulas in many columns
-            for(col++; col <= worksheet.Dimension.End.Column; col++)
-            {
-                cell = worksheet.Cells[endRow, col];
-
-                if (IsDataCell(cell))
-                {
-                    cell.FormulaR1C1 = GenerateFormula(worksheet, startRow, endRow, col);
-                }
-                else if(!IsEmptyCell(cell))
-                {
-                    return;
-                }
-            }
-
-
-
-        }
-
-
+        /* Some utility methods needed by the Formula generators */
 
 
         /// <summary>
@@ -234,7 +128,7 @@ namespace CompatableExcelCleaner
         /// </summary>
         /// <param name="cell">the cell being checked</param>
         /// <returns>true if the cell has no text and false otherwise</returns>
-        private static bool IsEmptyCell(ExcelRange cell)
+        internal static bool IsEmptyCell(ExcelRange cell)
         {
             return cell.Text == null || cell.Text.Length == 0;
         }
@@ -246,7 +140,7 @@ namespace CompatableExcelCleaner
         /// </summary>
         /// <param name="cell">the cell being checked</param>
         /// <returns>true if the cell contains a dollar value and false otherwise</returns>
-        private static bool IsDataCell(ExcelRange cell)
+        internal static bool IsDataCell(ExcelRange cell)
         {
             return cell.Text.StartsWith("$") || (cell.Text.StartsWith("($") && cell.Text.EndsWith(")"));
         }
@@ -261,7 +155,7 @@ namespace CompatableExcelCleaner
         /// <param name="endRow">the ending row of the formula range</param>
         /// <param name="col">the column the formula is for</param>
         /// <returns>the proper formula for the specified formula range</returns>
-        private static string GenerateFormula(ExcelWorksheet worksheet, int startRow, int endRow, int col)
+        internal static string GenerateFormula(ExcelWorksheet worksheet, int startRow, int endRow, int col)
         {
             ExcelRange cells = worksheet.Cells[startRow + 1, col, endRow - 1, col];
